requirements.txt
alembic==1.10.4
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.30.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
